refactor(runtime): quarantine legacy subagent graph shims
- Move active search, RAG, and tutor helper logic into native runtime modules while preserving graph.py compatibility shims
- Extract Code Studio scaffold primitive/kind mapping into a typed contract module
- Update tests and add a runtime cleanup audit for the remaining debt boundary

Verification:
- uv run --with pytest --with hypothesis --with pytest-asyncio pytest tests/unit/test_code_studio_template_scaffold.py tests/unit/test_subagent_phase3.py tests/unit/test_subagent_search.py tests/unit/test_sprint202_curated_cards.py -q --tb=short
- uv run --with pytest --with hypothesis --with pytest-asyncio pytest tests/unit/test_direct_tool_rounds_runtime.py -q --tb=short
- uv run --with ruff ruff check app/engine/multi_agent/code_studio_scaffold_contract.py app/engine/multi_agent/code_studio_template_scaffold.py app/engine/multi_agent/subagents/rag app/engine/multi_agent/subagents/search app/engine/multi_agent/subagents/tutor
- uv run --with ruff ruff check app/ --select=E9,F63,F7
- git diff --check
