+++ maritime-ai-service/app/engine/multi_agent/subagents/search/__init__.py
@@ -1,10 +1,13 @@
-"""Product search subgraph — parallel platform search via Send() map-reduce."""
+"""Product-search native helpers.
+
+LangGraph builders were removed; use ``runtime`` and ``workers`` helpers for
+WiiiRunner-native execution.
+"""
 
 from app.engine.multi_agent.subagents.search.state import (
     PlatformWorkerState,
     SearchSubgraphState,
 )
-# build_search_subgraph removed (De-LangGraphing Phase 3) — pipeline runs via WiiiRunner
 
 __all__ = [
     "PlatformWorkerState",
